US Cab Investment Analysis
The objective of this study was to analyze the operations of Pink Cab and Yellow Cab, two ride-hailing services operating in the United States. We performed several hypothesis tests and statistical analyses on a dataset of 359,392 rides to gain insights into the market and help investors make data-driven decisions.

Our analysis found that Yellow Cab dominates the market, covering 76.43% of trips compared to Pink Cab's 23.57%. We also found that distance traveled was positively correlated with the price charged and profit, suggesting that longer trips generate higher profits. Additionally, we found that cab usage is higher in larger cities and on certain days of the week and months of the year.

Regarding customer demographics, we found that age group does not have a significant impact on payment preferences, but there is a statistically significant association between age group and gender with respect to the number of cab service transactions.

Finally, we found that there is a relationship between the number of rides and US holidays, which could potentially be leveraged by cab services to generate more revenue during peak holiday seasons.

Based on our findings, we recommend that investors looking to invest in the ride-hailing sector consider Yellow Cab as a more profitable option due to its larger market share. Also, we suggest that investors pay close attention to peak demand times, such as holidays, as they could present opportunities for increased revenue